change: sdk DocumentManager, moved methods from DocumentManager (getCorrectName, getFilesInfo, createDemo)
